Changed SMR/methanol as a baseclass setup to be more clean
