+++ h2integrate/core/supported_models.py
@@ -11,11 +11,6 @@ from h2integrate.storage.hydrogen.tank_baseclass import (
     HydrogenTankPerformanceModel,
 )
 from h2integrate.converters.wind.wind_plant_pysam import PYSAMWindPlantPerformanceModel
-from h2integrate.converters.methanol.methanol_plant import (
-    MethanolPlantCostModel,
-    MethanolPlantFinanceModel,
-    MethanolPlantPerformanceModel,
-)
 from h2integrate.converters.wind.dummy_wind_turbine import DummyPlantCost, DummyPlantPerformance
 from h2integrate.converters.ammonia.ammonia_baseclass import (
     AmmoniaCostModel,
@@ -33,6 +28,11 @@ from h2integrate.converters.hydrogen.pem_electrolyzer import (
 from h2integrate.converters.hydrogen.dummy_electrolyzer import (
     DummyElectrolyzerCostModel,
     DummyElectrolyzerPerformanceModel,
+)
+from h2integrate.converters.methanol.smr_methanol_plant import (
+    SMRMethanolPlantCostModel,
+    SMRMethanolPlantFinanceModel,
+    SMRMethanolPlantPerformanceModel,
 )
 from h2integrate.converters.hydrogen.eco_tools_pem_electrolyzer import (
     ECOElectrolyzerCostModel,
@@ -63,9 +63,9 @@ supported_models = {
     "ammonia_cost": AmmoniaCostModel,
     "steel_performance": SteelPerformanceModel,
     "steel_cost": SteelCostAndFinancialModel,
-    "methanol_plant_performance": MethanolPlantPerformanceModel,
-    "methanol_plant_cost": MethanolPlantCostModel,
-    "methanol_plant_financial": MethanolPlantFinanceModel,
+    "smr_methanol_plant_performance": SMRMethanolPlantPerformanceModel,
+    "smr_methanol_plant_cost": SMRMethanolPlantCostModel,
+    "methanol_plant_financial": SMRMethanolPlantFinanceModel,
     # Transport
     "cable": CablePerformanceModel,
     "pipe": PipePerformanceModel,
